refactor(cxl-ui): replace cxl-marketing-nav component files, remove obsolete styles
