+++ packages/cxl-lumo-styles/scss/themes/vaadin-context-menu-list-box.scss
@@ -13,7 +13,7 @@
   }
 }
 
-:host([theme~="cxl-navigation"].vaadin-menu-list-box) {
+:host([theme~="cxl-marketing-nav"].vaadin-menu-list-box) {
   [part="items"]
     ::slotted(.section-header.vaadin-menu-item:not(.section-header):hover:not([disabled])) {
     /* revert default vaadin-context-menu-list-box hovering style, without losing other styles styles applied */
@@ -25,14 +25,6 @@
   }
 }
 
-@media screen and (min-width: 768px) and (min-height: 768px) {
-  :host([theme~="cxl-markting-nav"]) {
-    display: block;
-    width: var(--cxl-vaadin-context-menu-item-max-width);
-    padding: 0 var(--lumo-space-s);
-  }
-}
-
-:host([theme~="cxl-navigation"]) [part="items"] ::slotted(.vaadin-menu-item) {
+:host([theme~="cxl-marketing-nav"]) [part="items"] ::slotted(.vaadin-menu-item) {
   border-radius: 0;
 }
