+++ packages/cxl-lumo-styles/scss/themes/vaadin-tab.scss
@@ -1,12 +1,3 @@
-:host([theme~="cxl-marketing-nav"]) ::slotted(a[href]:hover) {
-  // text-decoration: underline !important; /* stylelint-disable-line declaration-no-important */
-  color: var(--lumo-primary-color) !important; /* stylelint-disable-line declaration-no-important */
-}
-
-:host([theme~="cxl-marketing-nav"][hidden]) {
-  visibility: hidden;
-}
-
 :host([theme~="cxl-featured-course-slider"]) {
   width: 100%;
   min-width: 100%;
