+++ packages/cxl-ui/scss/global/cxl-marketing-nav.scss
@@ -24,46 +24,6 @@ cxl-marketing-nav {
   }
 }
 
-#overlays-wrapper {
-  display: none;
-}
-
-@media screen and (min-width: 568px) {
-  #overlays-wrapper {
-    position: fixed;
-    z-index: 200;
-    display: flex;
-    background-color: var(--lumo-base-color);
-    border: 1px solid var(--lumo-base-color);
-    // Use top and bottom border prevent panel divider from reaching full height.
-    border-top-width: var(--lumo-space-m);
-    border-bottom-width: var(--lumo-space-m);
-    border-radius: 6px;
-    // Second shadow set to emulate border
-    box-shadow: 0 8px 24px 0 rgba(0, 0, 0, 0.25), 0 0 1px 1px #d2d5da;
-    opacity: 1;
-    transition: opacity 0.2s, left 0.2s;
-
-    vaadin-context-menu-overlay[theme~="cxl-marketing-nav"] {
-      display: flex;
-      visibility: visible;
-      background-color: var(--lumo-base-color);
-      opacity: 1; // show after being moved to the overlays-wrapper element
-      transition: all 0.25s;
-
-      + vaadin-context-menu-overlay[theme~="cxl-marketing-nav"] {
-        border-left: 1px solid var(--lumo-shade-20pct);
-      }
-    }
-
-    &[hidden] {
-      visibility: hidden;
-      opacity: 0;
-      transition: all 0.1s;
-    }
-  }
-}
-
 /**
  * PARTIALLY CORRECT
  * Apparently our element styles get hoisted also into `vaadin-overlay` shadow parts.
