+++ packages/cxl-ui/src/components/cxl-marketing-nav.js
@@ -1,159 +1,76 @@
 /* eslint-disable import/no-extraneous-dependencies */
-import { LitElement, html, render } from 'lit';
-import { customElement, property, query } from 'lit/decorators.js';
-import { unsafeHTML } from 'lit/directives/unsafe-html.js';
+import { LitElement, html } from 'lit';
+import { customElement, property, state, queryAll } from 'lit/decorators.js';
+import { MediaQueryController } from '@vaadin/component-base/src/media-query-controller.js';
 import '@conversionxl/cxl-lumo-styles';
 import { registerGlobalStyles } from '@conversionxl/cxl-lumo-styles/src/utils';
-import { MediaQueryController } from '@vaadin/component-base/src/media-query-controller.js';
-import { throttle } from 'lodash-es';
 import cxlMarketingNavStyles from '../styles/cxl-marketing-nav-css.js';
 import cxlMarketingNavGlobalStyles from '../styles/global/cxl-marketing-nav-css.js';
-import '@vaadin/button';
-import '@vaadin/tabs';
-import '@vaadin/context-menu';
+import '@vaadin/icon';
+import '@vaadin/menu-bar';
 
 @customElement('cxl-marketing-nav')
 export class CXLMarketingNavElement extends LitElement {
-  @property({ type: Object })
-  contextMenuItems;
-
-  @property({ type: Boolean, reflect: true })
-  fixed = false;
-
-  @query('.menu-items')
-  menuShadowItemsElement;
-
-  @property({ type: HTMLElement })
-  get menuItemSearchElement() {
-    return this.shadowRoot.querySelector('.menu-item-search');
-  }
-
-  @property({ type: NodeList })
-  get menuItemsElements() {
-    return this.querySelectorAll('.menu-items');
-  }
-
-  @property({ type: Number })
-  menuShadowItemsSelectedIdx = -1;
-
-  // Conditionally hide distracting shadow menu items. Used for checkout, etc.
-  @property({ type: Boolean, reflect: true })
-  minimal = false;
-
-  // MediaQueryController.
-  @property({ type: Boolean, reflect: true })
-  wide;
-
-  // Device Detector media query.
-  _wideMediaQuery = '(min-width: 568px)';
-
-  @property({ type: Boolean, reflect: true })
-  _phone;
-
-  // @see https://github.com/vaadin/web-components/blob/de3db720ec8448a26d2f84d00965a9e369a1c3fb/packages/select/src/vaadin-select.js#L297
-  _phoneMediaQuery = '(max-width: 568px), (max-height: 568px)';
-
-  @property({ type: HTMLElement })
-  submenuOverlay = null;
-
   static get styles() {
     return [cxlMarketingNavStyles];
   }
 
-  // eslint-disable-next-line class-methods-use-this
-  get overlaysWrapperElement() {
-    return document.body.querySelector('#overlays-wrapper');
+  _contextMenuItems = { global: [], primary: [] };
+
+  _phoneMediaQuery = '(max-width: 568px), (max-height: 568px)';
+
+  @queryAll('vaadin-menu-bar')
+  menuBars;
+
+  @property({ reflect: true, attribute: 'wide' }) wide = false;
+
+  @state()
+  groups = [];
+
+  @property({ type: Boolean, attribute: 'show-search' }) showSearch = false;
+
+  @property({ type: Object })
+  get contextMenuItems() {
+    return this._contextMenuItems;
   }
 
-  render() {
-    return html`
-      <vaadin-tabs
-        id="menu-shadow-items"
-        class="menu-items"
-        selected="${this.menuShadowItemsSelectedIdx}"
-        orientation="horizontal"
-        theme="cxl-marketing-nav hide-scroll-buttons minimal"
-        @selected-changed="${(e) => {
-          this.menuShadowItemsSelectedIdx = e.detail.value;
-        }}"
-        @click="${this._menuShadowItemsClick}"
-      >
-        <vaadin-tab class="menu-item menu-item-logo" theme="cxl-marketing-nav">
-          <a href="https://cxl.com"
-            ><vaadin-icon
-              icon="cxl:logo"
-              style="width: var(--lumo-icon-size-xl, 48px);"
-            ></vaadin-icon
-          ></a>
-        </vaadin-tab>
-        <vaadin-tab
-          class="menu-item menu-item-search"
-          theme="cxl-marketing-nav"
-          ?hidden="${this.minimal}"
-        >
-          <a
-            ><vaadin-icon icon="lumo:search"></vaadin-icon> Search
-            <vaadin-icon icon="lumo:dropdown"></vaadin-icon
-          ></a>
-          <vaadin-context-menu
-            close-on="outside-click"
-            open-on="click"
-            theme="cxl-marketing-nav"
-          ></vaadin-context-menu
-        ></vaadin-tab>
-        <vaadin-tab
-          class="menu-item menu-item-menu-toggle"
-          theme="cxl-marketing-nav"
-          ?hidden="${this.minimal}"
-          @click=${this._toggleMobileMenu}
-        >
-          <a
-            >Menu <vaadin-icon icon="lumo:menu"></vaadin-icon
-            ><vaadin-icon icon="lumo:cross"></vaadin-icon
-          ></a>
-        </vaadin-tab>
-      </vaadin-tabs>
-
-      <nav>
-        <slot @slotchange=${this._boundSetupSlottedMenuItems}></slot>
-      </nav>
-    `;
+  set contextMenuItems(data) {
+    this._contextMenuItems = { ...data };
+    const parseItem = (item) => {
+      const newItem = {
+        component: this.createItem(item),
+      };
+      if (item.children?.length) {
+        newItem.children = [{ component: 'back' }, ...item.children].map(parseItem);
+      }
+      return newItem;
+    };
+    const groups = [];
+    Object.keys(data).forEach((group) => {
+      const items = data[group];
+      groups.push({
+        name: group,
+        items: [...items?.map(parseItem)],
+      });
+    });
+    this.groups = groups;
+    this.requestUpdate('groups');
   }
 
   connectedCallback() {
     super.connectedCallback();
-
-    this.addController(
-      new MediaQueryController(this._wideMediaQuery, (matches) => {
-        if (this.wide !== undefined && this.wide !== matches) {
-          window.location.reload();
-        }
-
-        this.wide = matches;
-      })
-    );
-
     this.addController(
       new MediaQueryController(this._phoneMediaQuery, (matches) => {
-        this._phone = matches;
+        this.wide = !matches;
+        this.mobileGroups = [
+          { name: 'primary', items: this.groups.map((group) => group.items).flat(1) },
+        ];
       })
     );
-
-    this._boundOnOverlayOpen = this._onOverlayOpen.bind(this);
-    const overlaysWrapper = document.createElement('div');
-    overlaysWrapper.id = 'overlays-wrapper';
-    overlaysWrapper.toggleAttribute('hidden', true);
-    document.body.appendChild(overlaysWrapper);
-    this._boundSetupSlottedMenuItems = this._setupSlottedMenuItems.bind(this);
   }
 
-  disconnectedCallback() {
-    super.disconnectedCallback();
-    document.body.removeEventListener('vaadin-overlay-open', this._boundOnOverlayOpen);
-    document.body.removeChild(this.overlaysWrapperElement);
-  }
-
-  firstUpdated(changedProperties) {
+  // eslint-disable-next-line class-methods-use-this
+  firstUpdated() {
     /**
      * Global styles.
      */
@@ -161,476 +78,172 @@ export class CXLMarketingNavElement extends LitElement {
       moduleId: 'cxl-marketing-nav-global',
     });
 
-    /**
-     * Configure context menu trigger on main link click.
-     *
-     * @see https://github.com/vaadin/vaadin-context-menu/blob/v4.3.12/src/vaadin-context-menu.html#L172
-     * @see https://www.nngroup.com/articles/split-buttons-navigation/
-     */
-    this.querySelectorAll('.menu-item > vaadin-context-menu').forEach((contextMenu) => {
-      // eslint-disable-next-line no-param-reassign
-      contextMenu.listenOn = contextMenu.parentElement;
-
-      /**
-       * Back button custom event.
-       *
-       * @since 2022.11.14
-       * @see https://app.clickup.com/t/3rgnekt
-       * @type {string}
-       */
-      // eslint-disable-next-line no-param-reassign
-      contextMenu.closeOn = 'backBtnClose';
+    requestAnimationFrame(() => {
+      this._replaceMenuIcon();
     });
-
-    /**
-     * Listen for `vaadin-overlay-open` and remove the `hidden` attribute from description elements.
-     */
-    document.addEventListener('vaadin-overlay-open', (e) => {
-      if (e.target.tagName.toLowerCase() === 'vaadin-context-menu-overlay') {
-        const contextMenuListBox = e.target.getFirstChild();
-        const listItems = contextMenuListBox.items;
-
-        listItems.forEach((listItem) => {
-          const description = listItem.querySelector('.vaadin-context-menu-item--description');
-
-          if (description) {
-            description.hidden = false;
-          }
-        });
-      }
-    });
-
-    /**
-     * Search form integration.
-     */
-    this._setupSearchForm();
-
-    /**
-     * Decide `<vaadin-tabs>` initial orientation.
-     */
-    this._updatedWide();
-
-    /**
-     * Highlight current menu item.
-     */
-    this._highlightCurrentMenuItem();
-
-    /**
-     * Handle `<vaadin-context-menu-overlay>`.
-     * Check every 1000ms the scroll position.
-     */
-    window.addEventListener(
-      'scroll',
-      throttle(() => {
-        this._maybeCloseContextMenuOverlay();
-      }, 1000)
-    );
-
-    document.body.addEventListener('vaadin-overlay-open', this._boundOnOverlayOpen);
-
-    super.firstUpdated(changedProperties);
   }
 
-  /**
-   * Adjusts the target element's position to fit within the available width.
-   *
-   * @param {HTMLElement} target - The target element to adjust.
-   * @return {string|boolean} The adjusted position value in pixels or `false` if no adjustment is needed.
-   */
-  static _adjustToWidth(target) {
-    if (target.clientWidth + target.offsetLeft > visualViewport.width) {
-      return `${
-        target.offsetLeft - (target.clientWidth + target.offsetLeft - visualViewport.width)
-      }px`;
-    }
-    return false;
-  }
-
-  /*
-   * Setup listeners on nav items to handle overlays events.
-   */
-  _setupSlottedMenuItems() {
-    const navItems = this.querySelectorAll('vaadin-tab');
-    const overlaysWrapper = this.overlaysWrapperElement;
-
-    navItems.forEach((tab) =>
-      tab.addEventListener('click', (e) => {
-        const overlays = overlaysWrapper.querySelectorAll(
-          'vaadin-context-menu-overlay[theme="cxl-marketing-nav"]'
-        );
-        const parentOffset = tab.parentElement.getBoundingClientRect();
-        overlaysWrapper.style.top = `${parentOffset.y + 0.75 * tab.clientHeight}px`;
-        overlaysWrapper.style.left = `${e.clientX}px`;
-
-        [...overlays].forEach((overlay) => {
-          overlay.close();
-        });
-      })
-    );
-  }
-
-  _onOverlayClose(e) {
-    const overlay = e.target;
-    if (overlay.parentElement !== this.overlaysWrapperElement) return;
-
-    document.body.appendChild(overlay);
-    if (!this.overlaysWrapperElement.children.length) {
-      this.overlaysWrapperElement.toggleAttribute('hidden', true);
+  updated(changes) {
+    if (changes.has('groups') || changes.has('wide')) {
+      this._updateContextMenuItems();
+      this._replaceMenuIcon();
     }
   }
 
-  /**
-   * Store overlay when triggered for `vaadin-context-menu-item` sub menu.
-   *
-   * @since 2022.11.14
-   * @see https://app.clickup.com/t/3rgnekt
-   * @private
-   */
-  _onOverlayOpen(e) {
-    const overlay = e.target;
-    if (overlay.getAttribute('theme').indexOf('cxl-marketing-nav') === -1) return;
+  render() {
+    const groups = this.wide ? this.groups : this.mobileGroups;
 
-    overlay.addEventListener('opened-changed', this._onOverlayClose.bind(this));
-
-    if (window.matchMedia(this._wideMediaQuery).matches) {
-      const overlays = document.querySelectorAll(
-        'vaadin-context-menu-overlay[theme="cxl-marketing-nav"]'
-      );
-      const firstOverlay = overlays[0];
-
-      if (this.overlaysWrapperElement) {
-        if (overlays.length > 1) {
-          const heightOffset =
-            this.overlaysWrapperElement.style?.maxHeight || firstOverlay.offsetTop;
-          this.overlaysWrapperElement.style.maxHeight = `${
-            window.innerHeight - (heightOffset || 0)
-          }px`;
-        }
-
-        this.overlaysWrapperElement.appendChild(overlay);
-        this.overlaysWrapperElement.toggleAttribute('hidden', false);
-      }
-
-      // check if position needs adjusting
-      this.overlaysWrapperElement.style.left =
-        this.constructor._adjustToWidth(this.overlaysWrapperElement) ||
-        this.overlaysWrapperElement.style.left;
-    }
-
-    /**
-     * Non-wide context menu "Back" button.
-     */
-    const backBtn = overlay.querySelector('.context-menu-item-back-button');
-
-    if (!backBtn) {
-      return;
-    }
-
-    this.submenuOverlay = overlay;
+    return html`
+      ${groups.map((group) => {
+        const { name, items } = group;
+        return html`
+          <nav id="menu-${name}-items" ?minimal=${!this.wide}>
+            <div class="container">
+              ${name === 'global' || !this.wide
+                ? html`
+                    <vaadin-menu-bar-button theme="tertiary cxl-marketing-nav">
+                      <a href="https://cxl.com">
+                        <vaadin-icon
+                          icon="cxl:logo"
+                          style="width: var(--lumo-icon-size-xl, 48px);"
+                        ></vaadin-icon>
+                      </a>
+                    </vaadin-menu-bar-button>
+                  `
+                : ''}
+              ${!this.wide ? this._renderSearch() : ''}
+              <slot></slot>
+              <vaadin-menu-bar theme="tertiary cxl-marketing-nav" .items=${items}></vaadin-menu-bar>
+              <slot name="${name}-after"></slot>
+              <slot name="${name}-end"></slot>
+            </div>
+          </nav>
+        `;
+      })}
+    `;
   }
 
-  /**
-   * Listen to back button click event,
-   * close the relevant overlay.
-   *
-   * @since 2022.11.14
-   * @see https://app.clickup.com/t/3rgnekt
-   * @private
-   */
-  _onBackBtnClick() {
-    if (!this.submenuOverlay) {
-      return;
-    }
-
-    this.submenuOverlay.dispatchEvent(new CustomEvent('backBtnClose', { bubbles: true }));
-  }
-
-  updated(changedProps) {
-    if (changedProps.has('contextMenuItems')) {
-      this._updatedContextMenuItems();
-    }
-
-    if (changedProps.has('wide')) {
-      this._updatedWide();
-    }
-
-    super.updated(changedProps);
-  }
-
-  /**
-   * Add `<vaadin-context-menu>` element for back button.
-   */
   // eslint-disable-next-line class-methods-use-this
-  _addBackButtonItem(contextMenuItems) {
-    const backBtn = {
-      component: 'back',
-    };
-
-    // eslint-disable-next-line no-param-reassign
-    contextMenuItems = [backBtn, ...contextMenuItems];
-
-    return contextMenuItems;
+  _renderSearch() {
+    return html`
+      <vaadin-menu-bar-button theme="tertiary cxl-marketing-nav" class="search-button">
+        <a>
+          <vaadin-icon icon="lumo:search"></vaadin-icon> Search
+          <vaadin-icon icon="lumo:dropdown"></vaadin-icon>
+        </a>
+        <vaadin-context-menu close-on="outside-click" open-on="click" theme="cxl-marketing-nav">
+        </vaadin-context-menu>
+      </vaadin-menu-bar-button>
+    `;
   }
 
-  /**
-   * Create `<vaadin-context-menu>` elements, with children nesting support.
-   * Support `{ component: a }`.
-   *
-   * @private
-   * @see https://github.com/vaadin/vaadin-context-menu/issues/254
-   */
-  _createContextMenuItems(contextMenuItems) {
-    contextMenuItems.forEach((item, i, self) => {
-      if (item.children) {
-        // eslint-disable-next-line no-param-reassign
-        item.children = this._addBackButtonItem(item.children);
-        this._createContextMenuItems(item.children);
-      }
+  // eslint-disable-next-line class-methods-use-this
+  _onBackBtnClick(e) {
+    e.stopImmediatePropagation();
+    [...document.body.querySelectorAll('vaadin-context-menu-overlay')].at(-1).close();
+  }
 
-      const menuItem = document.createElement('vaadin-context-menu-item');
+  createItem({ text, description, sectionheader, component, icon, href, children, depth, split }) {
+    const item = document.createElement('vaadin-context-menu-item');
+    if (sectionheader) {
+      item.classList.add('section-header');
+    }
 
-      if (item.sectionheader) {
-        menuItem.classList.add('section-header');
+    if (split) {
+      item.classList.add('menu-item-split-nav');
+    }
 
-        const label = document.createTextNode(item.text);
-        menuItem.appendChild(label);
+    if (component === 'hr') {
+      return document.createElement('hr');
+    }
 
-        // eslint-disable-next-line no-param-reassign
-        self[i] = { component: menuItem };
-      } else if (item.component === 'a') {
-        const link = document.createElement('a');
+    if (icon) {
+      const vaadinIcon = document.createElement('vaadin-icon');
+      vaadinIcon.setAttribute('icon', `lumo:${icon}`);
+      item.appendChild(vaadinIcon);
+    }
 
-        if (item.href) {
-          link.href = item.href;
-          link.innerHTML = item.text;
+    if (href && !children && !sectionheader) {
+      const link = document.createElement('a');
+      link.href = href;
+      link.innerText = text;
+      item.appendChild(link);
+    } else if (text) {
+      const labelElement = sectionheader
+        ? document.createElement('h6')
+        : document.createElement('span');
+      labelElement.classList.add('vaadin-context-menu-item--label');
+      labelElement.appendChild(document.createTextNode(text));
+      item.appendChild(labelElement);
+    }
+
+    if (description) {
+      const descriptiondiv = document.createElement('div');
+      descriptiondiv.classList.add('vaadin-context-menu-item--description');
+      descriptiondiv.appendChild(document.createTextNode(description));
+      item.appendChild(descriptiondiv);
+    }
+
+    if (children?.length && depth === 0) {
+      const vaadinIcon = document.createElement('vaadin-icon');
+      vaadinIcon.setAttribute('icon', 'lumo:dropdown');
+      vaadinIcon.classList.add('vaadin-context-menu-item--icon');
+      item.appendChild(vaadinIcon);
+    }
+
+    if (component === 'back') {
+      const backBtn = document.createElement('vaadin-button');
+
+      backBtn.classList.add('context-menu-item-back-button');
+      backBtn.innerHTML = '<vaadin-icon icon="lumo:angle-left"></vaadin-icon> Back';
+
+      item.classList.add('back-button-menu-item');
+      item.appendChild(backBtn);
+
+      item.addEventListener('click', this._onBackBtnClick.bind(this));
+    }
+
+    return item;
+  }
+
+  _updateContextMenuItems() {
+    if (!this.menuBars) return;
+
+    [...this.menuBars].forEach((menu) => {
+      const splitItems = menu.shadowRoot.querySelectorAll(
+        'vaadin-context-menu-item.menu-item-split-nav'
+      );
+      [...splitItems].forEach((item) => {
+        if (!item.parentElement.classList.contains('menu-item-split-nav')) {
+          item.parentElement.classList.add('menu-item-split-nav');
         }
-
-        menuItem.appendChild(link);
-        // eslint-disable-next-line no-param-reassign
-        self[i] = { component: menuItem };
-      }
-
-      if (item.description) {
-        const descriptionItem = document.createElement('div');
-
-        descriptionItem.classList.add('vaadin-context-menu-item--description');
-
-        // Set to hidden, to calculate currently opened menu width and use it for description.
-        descriptionItem.hidden = true;
-
-        render(html`${unsafeHTML(item.description)}`, descriptionItem);
-
-        menuItem.appendChild(descriptionItem);
-        menuItem.classList.add('has-description');
-      }
-
-      if (item.component === 'back') {
-        const backBtn = document.createElement('vaadin-button');
-
-        backBtn.classList.add('context-menu-item-back-button');
-        backBtn.innerHTML = '<vaadin-icon icon="lumo:angle-left"></vaadin-icon> Back';
-
-        menuItem.classList.add('back-button-menu-item');
-        menuItem.appendChild(backBtn);
-
-        menuItem.addEventListener('click', this._onBackBtnClick.bind(this));
-
-        // eslint-disable-next-line no-param-reassign
-        self[i] = { component: menuItem };
-      }
-    });
-
-    return contextMenuItems;
-  }
-
-  /**
-   * Highlight menu item with special class.
-   * Improves visual "Where am I?" navigation clarity.
-   *
-   * @since 2020.04.12
-   * @private
-   */
-  _highlightCurrentMenuItem() {
-    this.menuItemsElements.forEach((menuItemsEl) => {
-      const currentMenuItemEl = menuItemsEl.querySelector('.current-menu-item');
-
-      if (currentMenuItemEl && currentMenuItemEl.id && menuItemsEl.items) {
-        const idx = menuItemsEl.items.findIndex((i) => i.id === currentMenuItemEl.id);
-
-        menuItemsEl.setAttribute('selected', idx);
-      }
-    });
-  }
-
-  /**
-   * Populate children `<vaadin-context-menu>` elements.
-   *
-   * @private
-   * @todo Links cannot be sub-menu hosts?
-   */
-  _updatedContextMenuItems() {
-    Object.values(this.contextMenuItems).forEach((items) => {
-      items.forEach((menuItem) => {
-        if (!menuItem.children || !menuItem.children.length) {
-          return;
-        }
-
-        const contextMenu = this.querySelector(
-          `vaadin-tab#menu-item-${menuItem.id} > vaadin-context-menu`
-        );
-
-        // Populate.
-        contextMenu.items = this._createContextMenuItems(menuItem.children);
-
-        // Prevent close on upstream events: clicks, keydown, etc
-        contextMenu.addEventListener('item-selected', (e) => {
-          e.stopImmediatePropagation();
-        });
       });
     });
   }
 
-  /**
-   * Re-orient menu items.
-   *
-   * @see https://github.com/vaadin/vaadin-context-menu/blob/v4.3.12/src/vaadin-device-detector.html#L12
-   * @see https://github.com/vaadin/vaadin-context-menu/issues/253
-   * @todo refactor w/ `<cxl-icon-nav>`?
-   */
-  _reorientMenuItems() {
-    let orientation = 'vertical';
+  _replaceMenuIcon() {
+    if (!this.menuBars) return;
 
-    if (this.wide) {
-      orientation = 'horizontal';
-    }
-
-    this.menuItemsElements.forEach((el) => {
-      el.setAttribute('orientation', orientation);
-      el.setAttribute('wide', this.wide);
-    });
-
-    this.menuShadowItemsElement.setAttribute('wide', this.wide);
-  }
-
-  /**
-   * Rotate search menu item context menu listener, because Chrome does not allow sending clicks to hidden elements.
-   *
-   * @see https://sookocheff.com/post/javascript/the-javascript-click-event-and-hidden-input-elements/
-   */
-  _rotateMenuItemSearchListenOn() {
-    let searchElement = this.menuItemSearchElement;
-
-    if (this.wide) {
-      searchElement = this.querySelector('.menu-item-search');
-    }
-
-    // Empty `cxl-marketing-nav` check.
-    if (searchElement) {
-      this.menuItemSearchElement.querySelector('vaadin-context-menu').listenOn = searchElement;
-    }
-  }
-
-  /**
-   * @private
-   */
-  _setupSearchForm() {
-    /**
-     * Configure `.menu-item-search`.
-     */
-    const menuItemSearchContextMenu =
-      this.menuItemSearchElement.querySelector('vaadin-context-menu');
-
-    /**
-     * `<vaadin-context-menu-item>` interferes with form input.
-     *
-     * @see https://github.com/vaadin/vaadin-item/blob/v2.1.1/src/vaadin-item-mixin.html#L136
-     */
-    menuItemSearchContextMenu.addEventListener(
-      'opened-changed',
-      (ee) => {
-        const searchForm = ee.target.$.overlay.querySelector('.search-form');
-
-        searchForm.addEventListener('keydown', (ef) => {
-          // Allow Esc.
-          if (ef.key !== 'Escape') {
-            ef.stopPropagation();
-          }
-        });
-      },
-      { once: true }
-    );
-
-    /**
-     * Avoid upstream default immediate close behavior.
-     */
-    menuItemSearchContextMenu.addEventListener('item-selected', (e) => {
-      e.stopImmediatePropagation();
-    });
-
-    /**
-     * Attach search form template.
-     */
-    const searchFormTemplate = this.querySelector('#cxl-marketing-nav-search-form-template') || '';
-
-    if (searchFormTemplate && 'content' in searchFormTemplate) {
-      menuItemSearchContextMenu.items = [
-        { component: searchFormTemplate.content.firstElementChild },
-      ];
-    }
-
-    /**
-     * Enable instant typing, avoid focus click.
-     */
-    menuItemSearchContextMenu.$.overlay.addEventListener('vaadin-overlay-open', (e) =>
-      e.target.querySelector('#search-input').focus()
-    );
-  }
-
-  /**
-   * @private
-   */
-  _updatedWide() {
-    this._reorientMenuItems();
-    this._rotateMenuItemSearchListenOn();
-  }
-
-  /**
-   * Mobile menu toggle.
-   *
-   * @see https://app.clickup.com/t/3phu3bv
-   * @since 2022.10.18
-   */
-  // eslint-disable-next-line class-methods-use-this
-  _toggleMobileMenu() {
-    const bodyElement = document.querySelector('body');
-
-    if (bodyElement) {
-      bodyElement.classList.toggle('cxl-mobile-menu-opened');
-    }
-  }
-
-  _menuShadowItemsClick(e) {
-    if (this.menuShadowItemsSelectedIdx !== -1) {
-      if (this.menuShadowItemsSelectedIdx === e.currentTarget.selected) {
-        e.currentTarget.selected = -1;
-        e.stopImmediatePropagation();
+    [...this.menuBars].forEach((menu) => {
+      const overflowMenuButton = menu.shadowRoot.querySelector(
+        'vaadin-menu-bar-button[part="overflow-button"]'
+      );
+      if (overflowMenuButton && !overflowMenuButton.iconFixed) {
+        const dots = overflowMenuButton.querySelector('.dots');
+        if (dots) {
+          dots.remove();
+        }
+        const menuIcon = document.createElement('vaadin-icon');
+        const closeIcon = document.createElement('vaadin-icon');
+        menuIcon.setAttribute('icon', 'lumo:menu');
+        closeIcon.setAttribute('icon', 'lumo:cross');
+        menuIcon.classList.add('menu-icon');
+        closeIcon.classList.add('close-icon');
+        overflowMenuButton.appendChild(closeIcon);
+        overflowMenuButton.appendChild(menuIcon);
+        overflowMenuButton.iconFixed = true;
       }
-    }
-  }
-
-  /**
-   * Maybe close context menu overlay on scroll?
-   * Only applicable if it's shown as dropdown.
-   *
-   * @see https://app.clickup.com/t/3q2z1a3
-   * @since 2022.10.19
-   */
-  _maybeCloseContextMenuOverlay() {
-    const el = document.querySelector('vaadin-context-menu-overlay');
-
-    if (el && !this._phone) {
-      el.close();
-    }
+    });
   }
 
   /**
